Refactor and expand tests for reply link and content
Moved and enhanced the test for reply link generation with Pleroma URLs from Test_Post to Test_Blocks, now using the render_reply_block method. Updated Test_Post to focus on testing the get_content method, ensuring it returns empty for specific block content. The filter_pleroma_object helper was also moved to Test_Blocks to support the new test location.
